# .specs/contracts/openapi.yaml
openapi: 3.0.4
info:
  title: Rampa Remittances API
  description: |
    API for Solana-based remittances app targeting Spain/Germany to LATAM corridor.
    
    ## Authentication Flow
    
    This API uses Web3Auth for user authentication with a token exchange pattern:
    
    1. **User Login**: User logs in via Web3Auth (Google, Apple, etc.) on the frontend
    2. **Token Exchange**: Frontend calls `/auth/web3auth/validate` with Web3Auth JWT token
    3. **API Token**: Backend validates Web3Auth token and returns our API JWT token
    4. **API Calls**: Frontend uses our API JWT token for all subsequent business API calls
    
    ## Security
    
    - All business endpoints require our API JWT token (Bearer authentication)
    - Web3Auth endpoints (`/auth/web3auth/*`) are public and don't require authentication
    - Health and inquiry endpoints are public
  version: 1.0.0
  contact:
    name: Rampa Team
    email: support@rampa.com
servers:
  - url: https://api.rampa.com
    description: Production server
  - url: https://staging-api.rampa.com
    description: Staging server
  - url: http://localhost:3001
    description: Development server
security:
  - BearerAuth: []
paths:
...
  /offramp/initiate:
    post:
      summary: Initiate crypto to fiat conversion
      tags: [OffRamp]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OffRampRequest'
      responses:
        '201':
          description: Off-ramp initiated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OffRampResponse'
        '400':
          description: Invalid request data
        '401':
          description: Unauthorized
        '403':
          description: Profile verification required for this operation
  /offramp:
    get:
      summary: Get all off-ramps
      tags: [OffRamp]
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, processing, completed, failed]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Off-ramps retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OffRampResponse'
        '401':
          description: Unauthorized
  /offramp/pending:
    get:
      summary: Get pending off-ramps
      tags: [OffRamp]
      responses:
        '200':
          description: Pending off-ramps retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OffRampResponse'
        '401':
          description: Unauthorized
  /offramp/stats/summary:
    get:
      summary: Get off-ramp statistics
      tags: [OffRamp]
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Off-ramp statistics retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalOffRamp:
                    type: number
                  totalFees:
                    type: number
                  completedOffRamp:
                    type: number
                  failedOffRamp:
                    type: number
        '401':
          description: Unauthorized
  /offramp/{id}:
    get:
      summary: Get specific off-ramp
      tags: [OffRamp]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Off-ramp retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OffRampResponse'
        '404':
          description: Off-ramp not found
        '401':
          description: Unauthorized

    post:
      summary: Process off-ramp
      tags: [OffRamp]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [providerTransactionId]
              properties:
                providerTransactionId:
                  type: string
      responses:
        '200':
          description: Off-ramp processed
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  status:
                    type: string
                  providerTransactionId:
                    type: string
                  completedAt:
                    type: string
                    format: date-time
        '400':
          description: Invalid request data
        '401':
          description: Unauthorized
  /offramp/{id}/process:
    post:
      summary: Process off-ramp
      tags: [OffRamp]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [providerTransactionId]
              properties:
                providerTransactionId:
                  type: string
      responses:
        '200':
          description: Off-ramp processed
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  status:
                    type: string
                  providerTransactionId:
                    type: string
                  completedAt:
                    type: string
                    format: date-time
        '400':
          description: Invalid request data
        '401':
          description: Unauthorized
  /offramp/{id}/fail:
    post:
      summary: Fail off-ramp
      tags: [OffRamp]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [failureReason]
              properties:
                failureReason:
                  type: string
      responses:
        '200':
          description: Off-ramp failed
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  status:
                    type: string
                  failureReason:
                    type: string
                  failedAt:
                    type: string
                    format: date-time
        '400':
          description: Invalid request data
        '401':
          description: Unauthorized
  /offramp/providers/{provider}/transaction/{providerTransactionId}:
    get:
      summary: Get off-ramp by provider
      tags: [OffRamp]
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
        - name: providerTransactionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Off-ramp retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OffRampResponse'
        '404':
          description: Off-ramp not found
        '401':
          description: Unauthorized
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Our API JWT token obtained from /auth/web3auth/validate
  schemas:
    TokenType:
      type: string
      enum: [USDC, EURC, SOL]
...
    OffRampRequest:
      type: object
      required: [amount, tokenType, bankAccount]
      properties:
        amount:
          type: number
          format: decimal
          minimum: 10
        tokenType:
          $ref: '#/components/schemas/TokenType'
        bankAccount:
          $ref: '#/components/schemas/BankAccount'
    BankAccount:
      type: object
      required: [iban, accountHolderName]
      properties:
        iban:
          type: string
        accountHolderName:
          type: string
        bankName:
          type: string
    OffRampResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        amount:
          type: number
          format: decimal
        fiatAmount:
          type: number
          format: decimal
        fiatCurrency:
          type: string
        tokenType:
          $ref: '#/components/schemas/TokenType'
        status:
          type: string
          enum: [pending, processing, completed, failed]
        exchangeRate:
          type: number
          format: decimal
        fee:
          type: number
          format: decimal
        estimatedDelivery:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
...